refactor team.py renamed team and player
